examples/medical-chatbot-agent/requirements.txt
flask
openai
python-dotenv
requests
